cleanup: remove redundant benchmark files
- Removed duplicate benchmark runners in favor of refactored run_all_benchmarks.py
- Removed simple memory test script as functionality is covered in main benchmarks
- Removed outdated RAM benchmark in favor of the optimized improved_ram_benchmark.py
- Removed redundant conflict test files
- Removed outdated visualization scripts that were consolidated

This commit streamlines the benchmark system to focus on the refactored implementation
while maintaining all benchmark functionality.
